Исправляет замечания наставника (частично)
