+++ src/view/page-header/site-menu-controls.js
