+++ src/view/page-header/trip-dates.js
