+++ src/view/page-header/trip-info.js
