+++ src/view/trip-event-form/city-option.js
