+++ src/view/trip-event-form/event-price.js
