+++ src/view/trip-event-form/form-photo.js
