+++ src/view/trip-event/empty-container.js
