+++ src/view/trip-event/trip-point.js
